Split test file into several
